fix a stepgen bug
Before this commit, the stepgen would sometimes fail to slow down all the
way when decelerating to a stop after completing a move in the negative
direction.  Even worse, it would then fail to notice that it was still
moving (vvvveeeeeerrrrryyyy slowly) in the negative direction.  It would
creep in the negative direction, at something like one step per minute,
but the .position-fb and .counts pins would not report the movement.
The steps were just lost.

This commit fixes an integer underflow bug, by casting the variables up
from u32 to s64 before doing the arithmetic.

+++ src/hal/drivers/mesa-hostmot2/stepgen.c
@@ -59,7 +59,7 @@ void hm2_stepgen_process_tram_read(hostmot2_t *hm2, long l_period_ns) {
         // representation of the current stepper position.
         // The fractional part gives accurate velocity at low speeds, and
         // sub-step position feedback (like sw stepgen).
-        acc_delta = acc - hm2->stepgen.instance[i].prev_accumulator;
+        acc_delta = (s64)acc - (s64)hm2->stepgen.instance[i].prev_accumulator;
         if (acc_delta > INT32_MAX) {
             acc_delta -= UINT32_MAX;
         } else if (acc_delta < INT32_MIN) {
